Resolve 64-bit warnings
gcc 4.7.2, on Sun, with 64-bit targets, raises a number of
warnings that are not as visible on other platforms.  This
patch for the 'bslstl_HashTable' test driver should shave
around 7,000 lines of warning-related spew from the log
files of our nightly tests.

+++ groups/bsl/bslstl/bslstl_hashtable.h
@@ -2230,7 +2230,8 @@ template <class KEY_CONFIG, class HASHER, class COMPARATOR, class ALLOCATOR>
 inline
 float HashTable<KEY_CONFIG, HASHER, COMPARATOR, ALLOCATOR>::loadFactor() const
 {
-    return static_cast<float>((double)size() / this->numBuckets());
+    return static_cast<float>(static_cast<double>(size())
+                                    / static_cast<double>(this->numBuckets()));
 }
 
 template <class KEY_CONFIG, class HASHER, class COMPARATOR, class ALLOCATOR>
